+++ test/unit/ResultSet/HydratingResultSetTest.php
@@ -75,7 +75,7 @@ class HydratingResultSetTest extends TestCase
     /**
      * @covers \Laminas\Db\ResultSet\HydratingResultSet::current
      */
-    public function testCurrent()
+    public function testCurrentHasData()
     {
         $hydratingRs = new HydratingResultSet;
         $hydratingRs->initialize([
@@ -83,6 +83,17 @@ class HydratingResultSetTest extends TestCase
         ]);
         $obj = $hydratingRs->current();
         self::assertInstanceOf('ArrayObject', $obj);
+    }
+
+    /**
+     * @covers \Laminas\Db\ResultSet\HydratingResultSet::current
+     */
+    public function testCurrentDoesnotHasData()
+    {
+        $hydratingRs = new HydratingResultSet;
+        $hydratingRs->initialize([]);
+        $result = $hydratingRs->current();
+        self::assertNull($result);
     }
 
     /**